requirements.txt
arrow==1.1.0
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.3
django-admin-tools==0.9.1
django-field-history==0.8.0
django-telegram-login==0.2.3
idna==2.10
install==1.3.4
Jinja2==3.0.1
jinja2-time==0.2.0
make==0.1.6.post2
MarkupSafe==2.0.1
oauthlib==3.1.0
phonenumbers==8.12.23
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
pyTelegramBotAPI==3.7.9
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.4
